requirements.txt
aiofiles==24.1.0
aiogram==3.22.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
certifi==2025.10.5
click==8.3.0
colorama==0.4.6
fastapi==0.118.1
frozenlist==1.8.0
h11==0.16.0
idna==3.10
loguru==0.7.3
magic-filter==1.0.12
multidict==6.7.0
pillow==11.3.0
propcache==0.4.0
pydantic==2.11.10
pydantic_core==2.33.2
python-dotenv==1.1.1
qrcode==8.2
sniffio==1.3.1
starlette==0.48.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.37.0
win32_setctime==1.2.0
yarl==1.22.0
